:: Created Take Operation and a few more things {-}

+++ ProjectExchange.API/Services/Product/Application/Models/ProductVm.cs
@@ -3,7 +3,7 @@ namespace Application.Models {
         public string Id { get; set; }
         public string Name { get; set; }
         public string ImgUrl { get; set; }
-        public int Weight { get; set; }
-        public int UnitPrice { get; set; }
+        public double Weight { get; set; }
+        public double UnitPrice { get; set; }
     }
 }
